+++ internal/consumer/event/event.go
@@ -1,4 +1,4 @@
-package eventConsumer
+package event
 
 import (
 	"fmt"
