+++ internal/events/processor/telegram.go
@@ -1,4 +1,4 @@
-package eventProcessor
+package processor
 
 import (
 	"errors"
